+++ lib/transport/CMakeLists.txt
@@ -12,6 +12,7 @@ set(TRANSPORT_HEADERS
     transport/multitransport.h
     transport/transport-factory-tls.h
     transport/transport-factory-socket.h
+    transport/tls-context.h
     PARENT_SCOPE)
 
 set(TRANSPORT_SOURCES
@@ -27,6 +28,7 @@ set(TRANSPORT_SOURCES
     transport/multitransport.c
     transport/transport-factory-tls.c
     transport/transport-factory-socket.c
+    transport/tls-context.c
     PARENT_SCOPE)
 
 add_test_subdirectory(tests)
