+++ lib/transport/tls-context.c
@@ -21,7 +21,7 @@
  * COPYING for details.
  */
 
-#include "tlscontext.h"
+#include "transport/tls-context.h"
 #include "str-utils.h"
 #include "messages.h"
 #include "compat/openssl_support.h"
