+++ lib/transport/transport-factory-tls.h
@@ -26,7 +26,7 @@
 #define TRANSPORT_FACTORY_TLS_H_INCLUDED
 
 #include "transport/transport-factory.h"
-#include "tlscontext.h"
+#include "transport/tls-context.h"
 
 typedef struct _TransportFactoryTLS TransportFactoryTLS;
