+++ modules/afsocket/afinet-source.h
@@ -26,7 +26,7 @@
 
 #include "afinet.h"
 #include "afsocket-source.h"
-#include "tlscontext.h"
+#include "transport/tls-context.h"
 
 typedef struct _AFInetSourceDriver
 {
