+++ modules/afsocket/afsocket-grammar.ym
@@ -47,7 +47,7 @@
 #include "systemd-syslog-source.h"
 #include "afsocket-systemd-override.h"
 
-#include "tlscontext.h"
+#include "transport/tls-context.h"
 
 
 static SocketOptions *last_sock_options;
